deeptools_qc requires alignment, so make that explicit. Also ensure alignment-free doesn't accidentally pull in alignment. Update docs

+++ snakePipes/common_functions.py
@@ -397,6 +397,10 @@ def runAndCleanup(args, cmd, logfile_name, temp_path):
             subprocess.call(["pkill", "-SIGTERM", "-P", str(p.pid)])
             print("SIGTERM sent to PID:", p.pid)
 
+    # Exit with an error if snakemake encountered an error
+    if p.returncode != 0:
+        sys.exit(p.returncode)
+
     # remove temp dir
     if (temp_path != "" and os.path.exists(temp_path)):
         shutil.rmtree(temp_path, ignore_errors=True)
